adicionado grafico e adicionado algumas logicas.

+++ gpio/gpio.h
@@ -1,7 +1,7 @@
 #ifndef __GPIO__
 #define __GPIO__
 
-void gpio_liga(int pino, double pwm);
+void gpio_liga(int pino, int pwm);
 void gpio_desliga(int pino);
 
 #endif
